Madhav Chandrasekher's Insight project for the summer 2020 session.
EtsyEdge: A Predictive Tool for Sellers on the Etsy platform  

EtsyEdge is a web application which takes a seller-product pair as  an input and uses descriptive statistics and benchmark ML algorithms to offer growth recommendations. The files required to run EtsyEdge are in the Projects folder and are described as follows:  DATA FILES  1. streamlit_candles.csv 2. streamlit_earrings.csv 3.